backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
httpx>=0.28.0
kokoro>=0.9
soundfile>=0.13.0
numpy>=2.0
faster-whisper>=1.1.0
edge-tts>=7.0.0
pyyaml>=6.0
python-multipart>=0.0.18
